Introduce API to configure the Environment at runtime
Before this commit the Environment was configured via config files,
which had to be present at the classpath root or the working directory.

There are use cases in which users want to configure the Environment
programmatically or want to reconfigure the Environment during runtime.
Example: to set a different "allowed difference" in various runs. This
would only have been possible by writing a config file and let
pdfcompare parse that config file again.

Now it is possible to call
  pdfComparator.setEnvironment(Environment)

The default interface implementations are ConfigFileEnvironment (which
parses a HOCON file) and SimpleEnvironment (with simple setter methods).

If no Environment is provided, the old mechanism is used as fallback.

Resolves #32